agent/src/requirements.txt
pywebview>=4.4

watchdog

pystray

pillow
